move source LuaMemory to lua_ext_gears
